requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-WTF==1.1.1
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
openpyxl==3.1.2
requests==2.31.0
